+++ src/common/util/token-counter.ts
@@ -1,7 +1,6 @@
 import { encoding_for_model, get_encoding, Tiktoken, TiktokenModel } from '@dqbd/tiktoken';
 
-import { findLLMOrThrow } from '~/modules/llms/llm.client';
-import { DLLMId, useModelsStore } from '~/modules/llms/store-llms';
+import { DLLMId, findLLMOrThrow, useModelsStore } from '~/modules/llms/store-llms';
 
 
 // Do not set this to true in production, it's very verbose
